Adding the standalone generation of JSON
